Rename class per review feedback
