+++ game-core/src/main/java/games/strategy/triplea/delegate/power/calculator/AaDefenseCombatValue.java
@@ -18,11 +18,11 @@ import lombok.Value;
 @Builder
 @Value
 @Getter(AccessLevel.NONE)
-class AaDefenseCalculator implements OffenseOrDefenseCalculator {
+class AaDefenseCombatValue implements CombatValue {
 
   @NonNull GameData data;
-  @NonNull AvailableSupportTracker friendlySupportTracker;
-  @NonNull AvailableSupportTracker enemySupportTracker;
+  @NonNull AvailableSupports friendlySupportTracker;
+  @NonNull AvailableSupports enemySupportTracker;
 
   @Override
   public StrengthOrRollCalculator getRoll() {
@@ -46,8 +46,7 @@ class AaDefenseCalculator implements OffenseOrDefenseCalculator {
 
   static class AaDefenseRoll extends StrengthOrRollCalculator {
 
-    AaDefenseRoll(
-        final AvailableSupportTracker friendlySupport, final AvailableSupportTracker enemySupport) {
+    AaDefenseRoll(final AvailableSupports friendlySupport, final AvailableSupports enemySupport) {
       super(friendlySupport, enemySupport);
     }
 
@@ -72,8 +71,8 @@ class AaDefenseCalculator implements OffenseOrDefenseCalculator {
 
     AaDefenseStrength(
         final GameData gameData,
-        final AvailableSupportTracker friendlySupport,
-        final AvailableSupportTracker enemySupport) {
+        final AvailableSupports friendlySupport,
+        final AvailableSupports enemySupport) {
       super(friendlySupport, enemySupport);
       this.gameData = gameData;
     }
